CT/API/MCP/requirements.txt
fastapi>=0.105.0
uvicorn>=0.23.2
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6